dev to main

+++ src/main/java/com/kroum/kroum/dto/request/ReviewCreateRequestDto.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 public class ReviewCreateRequestDto {
 
     @Schema(description = "별점, 1 ~ 5 사이의 정수", example = "3")
-    private double rating;
+    private int rating;
 
     @Schema(description = "리뷰 텍스트", example = "여기 진짜 ㄹㅇ 좋아요")
     private String content;
